Move Noop Metrics implementation to one class.
